Scope Camera for Sniper Rifles
No more lazy scope overlays
Used the Optiwand code to create a "camera" for the rifle's scope.

It's necessary to have a separate UV map for the lens , and the UV MAP  IT MUST BE TAKE CORRECTLY IN FRONT OF THE LENS AND RESPECTING THE CORRECT VERTICAL ORIENTATION!
Otherwise the camera will move in different directions in respect of the gun movement.